FIX : 가게 상세페이지 feed 가져오기 서버 연결 수정
