+++ src/custom/jh/useGetShopDetail.ts
@@ -2,7 +2,21 @@ import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { queryKeys } from '../../apis/queries';
 import api from '../../shared/api';
 import { apiPath } from '../../shared/path';
-import { IFeedList } from '../../components/FeedContents2';
+import { getToken } from '../../apis/getToken';
+
+export interface IFeedList {
+  nickname: string;
+  profilePic : string | undefined | null;
+  createdAt : string | Date;
+  feedPic : string;
+  comment : string | null;
+  tags : [] | string[];
+  shopId : number;
+  shopName : string;
+  shopAddress : string;
+  shopThumbnail : string;
+  isScrap : boolean;
+};
 
 export const useGetShopDetail = (param: number | undefined) => {
   const queryClinet = useQueryClient();
@@ -27,11 +41,17 @@ export const useGetShopDetail = (param: number | undefined) => {
 };
 
 export const useGetShopDetailFeed = (param: number | undefined) => {
-  const queryClinet = useQueryClient();
+  const token = getToken();
+  // const queryClinet = useQueryClient();
   const { data, mutate, isLoading, isError } = useMutation({
     mutationKey: queryKeys.GET_SHOP_DETAIL_FEED,
     mutationFn: async () => {
-      const {data}: {data: IFeedList[]} = await api.get(`${apiPath.toShopDetail}/${param}/feed`);
+      console.log('실행됐나?');
+      const {data} = await api.get(`${apiPath.toShopDetail}/${param}/feed2`, {
+        headers: {
+          authorization: `${token}`,
+        },
+      });
       console.log('피드 데이터', data);
       return data;
     },
